requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.51.0
Pillow>=10.0.0

# Evaluation
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data preparation
opencv-python>=4.8.0
tqdm>=4.65.0

# HuggingFace dataset download
huggingface_hub>=0.20.0

# Gemini API evaluation (optional)
# google-genai>=1.0.0
